+++ src/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h1>
+        Hello
+      </h1>
+    </div>
+  );
+}
+
+export default App;
